add setting and checkout interface
