+++ app/src/main/res/menu/home_menu.xml
@@ -1,0 +1,21 @@
+<?xml version="1.0" encoding="utf-8"?>
+<menu xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto">
+
+    <item
+        android:id="@+id/add"
+        android:icon="@mipmap/ic_add"
+        android:title="Add new Child Account"
+        app:showAsAction="ifRoom" />
+
+    <item
+
+        android:id="@+id/setting"
+        android:title="Settings"
+        app:showAsAction="withText" />
+    <item
+        android:id="@+id/logout"
+        android:title="Logout"
+        app:showAsAction="withText" />
+
+</menu>
